# Conflicts:
#	src/winds/CMakeLists.txt
#	src/winds/Solver_GPU_GlobalMemory.cu
#	src/winds/Solver_GPU_GlobalMemory.h

+++ src/plume/GLE_Solver.cpp
@@ -93,7 +93,7 @@ void GLE_Solver_CPU::solve(Particle *p, double &par_dt, TURBGeneralData *TGD, PL
                   static_cast<float>(PGD->threadRNG[omp_get_thread_num()]->norRan()),
                   static_cast<float>(PGD->threadRNG[omp_get_thread_num()]->norRan()) };
 #else
-  vec3 vRandn = { PGD->RNG->norRan(), PGD->RNG->norRan(), PGD->RNG->norRan() };
+  vec3 vRandn = { (float)PGD->RNG->norRan(), (float)PGD->RNG->norRan(), (float)PGD->RNG->norRan() };
 #endif
 
   // now calculate a bunch of values for the current particle
