+++ src/winds/WindProfilerWRF.h
@@ -41,7 +41,7 @@
 class WindProfilerWRF : public WindProfilerType
 {
 private:
-  void interpolate(float &, float &, int &, std::vector<float> &, std::vector<float> &, int, int);
+  void interpolate(float &, float &, long &, std::vector<float> &, std::vector<float> &, int, int);
 
 public:
   WindProfilerWRF()
